// Source/WebSPA/ClientApp/src/app/shared/models/localization.model.ts
export interface Localization {
    employee: string;
    employees: string; 
    login: string;
    organization: string; 
    department: string;  
    position: string;  
    equipment: string;  
    type: string; 
    status: string;  
    place: string;  
    inventoryNumber: string;  
    serialNumber: string;  
    equipmentNote: string;  
    dateOfCreation: string;  
    dateOfInstallation: string;  
    disposal: string;
    disposalName: string;
    disposalDate: string;
    disposalResalePrice: string;  
    disposalNote: string;  
    address: string;  
    writeOff: string;
    writeOffName: string;
    writeOffDate: string;
    writeOffLiquidationValue: string;  
    writeOffNote: string;  
    apply: string;
    main: string;
    add: string;
    delete: string;
    edit: string;
    search: string;
    close: string;
    username: string;
    password: string;
    oldPassword: string;
    newPassword: string;
    email: string;
    register: string;
    authorization: string;
    forgotYourPassword: string;
    changePassword: string;
    passwordConfirmation: string;
    confirmEmail: string;
    passwordRecovery: string;
    checkEmailAfterRegister: string;
    checkEmailAfterResetPassword: string;
    emailConfirmed: string;
    emailConfirmationError: string;
    passwordChanged: string;
    tableSetup: string;
    availableFields: string;
    selectedFields: string;
    attachFile: string;
    software: string;
    softwareNote: string;
    developer: string;
    license: string;
    licenseCount: string;
    licenseType: string;
    licenseCode: string;
    lisenseNote: string;
    usedLicenses: string;
    beforeDeletingAletMessage: string;
    loadingDataAletMessage: string;
    navigationMenu: string;
    userMenu: string;
    returnToStartPage: string;
    profile: string;
    exit: string;
    wrongLoginOrPassword: string;
    firstName: string;
    lastName: string;
    patronymicName: string;
    users: string;
    phoneNumber: string;
    roles: string;
    installedSoftwares: string;
    availableSoftwares: string;
    installedOnEquipment: string,
    availableEquipment: string;
    administration: string;
    signIn: string;
    file: string;
    user: string;
    userRoles: string;
    aboutMe: string;
    change: string;
    equipmentTypes: string;
    equipmentStatuses: string;
    organizations: string;
    offices: string;
    confirmDelete: string;
    cancel: string;
    loading: string;
    entitiesNotFound: string;
    fieldNotFilled: string;
}